% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LXvolcan.R
\name{LXvolcan}
\alias{LXvolcan}
\title{LXvolcan}
\usage{
LXvolcan(data_type, data_file, FC)
}
\arguments{
\item{FC}{}
}
\description{LXvolcan can be used to analyze the differently expressed genes or metabolites.}
\examples{

install.packages("devtools")

library(devtools)

install_github("gluck4668/LXvolcan")

library(LXvolcan)

#--------------------------------
data(gene_data_example)
data(meta_data_example)
#--------------------------------

rm(list=ls())

if(!is.null(dev.list()))
  dev.off()

devtools::load_all()

data_type="gene" # it should be "gene" or "metabolite"

data_file = "gene_data.xlsx"

FC = 2 #(FC is the Fold Change and should be 2 or NULL )

LXvolcan (data_type,data_file,FC)

}
